+++ apps/astarte_housekeeping_api/lib/astarte_housekeeping_api/application.ex
@@ -23,6 +23,8 @@ defmodule Astarte.Housekeeping.API.Application do
   alias Astarte.Housekeeping.API.Config
   alias Astarte.RPC.Config, as: RPCConfig
 
+  @app_version Mix.Project.config()[:version]
+
   # See http://elixir-lang.org/docs/stable/elixir/Application.html
   # for more information on OTP Applications
   def start(_type, _args) do
@@ -32,7 +34,7 @@ defmodule Astarte.Housekeeping.API.Application do
       {&:logger_filters.domain/2, {:stop, :equal, [:progress]}}
     )
 
-    Logger.info("Starting application.", tag: "housekeeping_api_start")
+    Logger.info("Starting application v#{@app_version}.", tag: "housekeeping_api_start")
 
     Config.validate!()
     RPCConfig.validate!()